Component3/requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.28.3
setuptools==75.5.0
pytest==7.2.2
